requirements.txt
google-generativeai
openai
requests
Pillow
google-cloud-translate
moondream
python-dotenv